+++ Clients/Models/__init__.py
@@ -6,3 +6,4 @@ from .SWUCardSearchConfiguration import SWUCardSearchConfiguration
 from .SWUTradingCard import SWUTradingCard
 from .SWUTradingCardBackedLocalCardResource import SWUTradingCardBackedLocalCardResource
 from .SWUTradingCardModelMapper import SWUTradingCardModelMapper
+from .DeckListImageGeneratorStyles import DeckListImageGeneratorStyles