+++ Clients/Utility/DraftListImageGenerator.py
@@ -4,43 +4,24 @@ from enum import Enum
 from PIL import Image
 from PIL.ImageFile import ImageFile
 import sys
-from ..Models import SWUTradingCardBackedLocalCardResource, ParsedDeckList, ParsedDeckListProviding
-
+from ..Models import SWUTradingCardBackedLocalCardResource, ParsedDeckList, ParsedDeckListProviding, DeckListImageGeneratorStyles
+from ..Config.SWUAppConfiguration import SWUAppConfigurationManager
 
 class DraftListImageGenerator:
-
-    STACKED_CARD_REVEAL_PERCENTAGE = 0.15
-    SIDEBOARD_LEFT_SPACING = 80
-    MAIN_DECK_COLUMN_SPACING = 10
-    MAIN_DECK_ROWS_SPACING = 40
-    LEADER_BASE_SPACING = 10
-
     class Option(str, Enum):
+        # TODO put it in to config
         COST_CURVE_LEADER_BASE_HORIZONTAL = "Cost curve - top, horizontal leader/base"
         COST_CURVE_LEADER_BASE_VERTICAL = "Cost curve - left, vertical leader/base"
-        # alphabet order option
-        # hide sideboard option
-        # spacing between general /row  COLS: 10, ROWS: 40
     
-    def __init__(self, parsed_deck_provider: ParsedDeckListProviding):
+    def __init__(self, parsed_deck_provider: ParsedDeckListProviding, 
+                 configuration_manager: SWUAppConfigurationManager):
+        self._configuration_manager = configuration_manager
         self._parsed_deck_list_provider = parsed_deck_provider
-        self._is_sideboard_enabled = True
-        self._is_alphabetical = False
         self._option = self.Option.COST_CURVE_LEADER_BASE_VERTICAL
 
     @property
-    def is_sideboard_enabled(self) -> bool:
-        return self._is_sideboard_enabled
-
-    def set_is_sideboard_enabled(self, is_enabled: bool):
-        self._is_sideboard_enabled = is_enabled
-
-    @property
-    def is_alphabetical(self) -> bool:
-        return self._is_alphabetical
-    
-    def set_is_alphabetical(self, value: bool):
-        self._is_alphabetical = value
+    def _deck_list_image_generator_styles(self) -> DeckListImageGeneratorStyles:
+        return self._configuration_manager.configuration.deck_list_image_generator_styles
 
     def set_option(self, option: 'DraftListImageGenerator.Option'):
         self._option = option
@@ -56,7 +37,6 @@ class DraftListImageGenerator:
             return self.generate_cost_curve_with_vertical_leader_base()
         
     # MARK: - cost curve
-    # sideboard images (optional)
     def _create_transparent_image(self, width: int, height: int) -> Image.Image:
         return Image.new('RGBA', (width, height), (255, 255, 255, 0))
     
@@ -71,14 +51,14 @@ class DraftListImageGenerator:
             
         width = 0
         for img in leader_and_base_image_files:
-            width += uniform_card_width + self.LEADER_BASE_SPACING
+            width += uniform_card_width + self._deck_list_image_generator_styles.leader_base_spacing_between
         combined_image_leader = self._create_transparent_image(width, uniform_card_height)
         offset = 0
         for img in leader_and_base_image_files:
             scaled_image = img.copy().convert('RGBA')
             scaled_image.thumbnail((uniform_card_width, uniform_card_width), Image.Resampling.BICUBIC)
             combined_image_leader.paste(scaled_image, (offset, 0), scaled_image)
-            offset += scaled_image.width + self.LEADER_BASE_SPACING
+            offset += scaled_image.width + self._deck_list_image_generator_styles.leader_base_spacing_between
         return combined_image_leader
     
     def _create_leader_base_stacked_vertical_image(self,
@@ -92,14 +72,14 @@ class DraftListImageGenerator:
             
         height = 0
         for img in leader_and_base_image_files:
-            height += uniform_card_height + self.LEADER_BASE_SPACING
+            height += uniform_card_height + self._deck_list_image_generator_styles.leader_base_spacing_between
         combined_image_leader = self._create_transparent_image(uniform_card_width, height)
         offset = 0
         for img in leader_and_base_image_files:
             scaled_image = img.copy().convert('RGBA')
             scaled_image.thumbnail((uniform_card_width, uniform_card_width), Image.Resampling.BICUBIC)
             combined_image_leader.paste(scaled_image, (0, offset), scaled_image)
-            offset += scaled_image.height + self.LEADER_BASE_SPACING
+            offset += scaled_image.height + self._deck_list_image_generator_styles.leader_base_spacing_between
         return combined_image_leader
     
     def create_cost_stack_col(self, images: List[ImageFile], uniform_card_width: int, uniform_card_height: int) -> Image.Image:
@@ -112,7 +92,7 @@ class DraftListImageGenerator:
                     height += uniform_card_height
                 else:
                     # obscure bottom cards
-                    height += int(uniform_card_height * self.STACKED_CARD_REVEAL_PERCENTAGE)
+                    height += int(uniform_card_height * self._deck_list_image_generator_styles.stacked_card_reveal_percentage)
             combined_image = self._create_transparent_image(uniform_card_width, height)
             
             curr_y = 0
@@ -120,7 +100,7 @@ class DraftListImageGenerator:
                 scaled_image = img.copy().convert('RGBA')
                 scaled_image.thumbnail((uniform_card_height, uniform_card_height), Image.Resampling.BICUBIC)
                 combined_image.paste(scaled_image, (0, curr_y), scaled_image)
-                curr_y += int(scaled_image.height * self.STACKED_CARD_REVEAL_PERCENTAGE)
+                curr_y += int(scaled_image.height * self._deck_list_image_generator_styles.stacked_card_reveal_percentage)
             
             return combined_image
 
@@ -160,13 +140,13 @@ class DraftListImageGenerator:
             width = 0
             for i in images:
                 height = max(height, i.height)
-                width += i.width + self.MAIN_DECK_COLUMN_SPACING
+                width += i.width + self._deck_list_image_generator_styles.main_deck_column_spacing
             combined_image = self._create_transparent_image(width, height)
             
             curr_x = 0
             for i, img in enumerate(images):
                 combined_image.paste(img, (curr_x, 0), img)
-                curr_x += img.width + self.MAIN_DECK_COLUMN_SPACING
+                curr_x += img.width + self._deck_list_image_generator_styles.main_deck_column_spacing
             
             return combined_image
                 
@@ -175,20 +155,20 @@ class DraftListImageGenerator:
             height = 0
             width = 0
             for i in images:
-                height += i.height + self.MAIN_DECK_ROWS_SPACING
+                height += i.height + self._deck_list_image_generator_styles.main_deck_row_spacing
                 width = max(width, i.width)
             combined_image = self._create_transparent_image(width, height)
             
             curr_y = 0
             for i, img in enumerate(images):
                 combined_image.paste(img, (0, curr_y), img)
-                curr_y += img.height + self.MAIN_DECK_ROWS_SPACING
+                curr_y += img.height + self._deck_list_image_generator_styles.main_deck_row_spacing
                 
             return combined_image
         
         main_deck_image = stack_rows([
-            create_card_type_row(lambda x: self._parsed_deck_list.all_units_with_cost(x, self._is_alphabetical), lambda x: self._parsed_deck_list.main_deck_with_cost(x)),
-            create_card_type_row(lambda x: self._parsed_deck_list.all_upgrades_and_events_with_cost(x, self._is_alphabetical), lambda x: self._parsed_deck_list.main_deck_with_cost(x)),
+            create_card_type_row(lambda x: self._parsed_deck_list.all_units_with_cost(x, self._deck_list_image_generator_styles.is_sorted_alphabetically), lambda x: self._parsed_deck_list.main_deck_with_cost(x)),
+            create_card_type_row(lambda x: self._parsed_deck_list.all_upgrades_and_events_with_cost(x, self._deck_list_image_generator_styles.is_sorted_alphabetically), lambda x: self._parsed_deck_list.main_deck_with_cost(x)),
         ])
         
         return main_deck_image
@@ -217,21 +197,21 @@ class DraftListImageGenerator:
         main_deck_image = self._create_main_deck_cost_curve_image(uniform_card_width, uniform_card_height)
         leader_base_image = self._create_leader_base_stacked_vertical_image(uniform_card_height, uniform_card_width)
 
-        leader_base_main_deck_image = self._create_transparent_image(main_deck_image.width + leader_base_image.width + self.MAIN_DECK_COLUMN_SPACING, 
+        leader_base_main_deck_image = self._create_transparent_image(main_deck_image.width + leader_base_image.width + self._deck_list_image_generator_styles.leader_base_spacing_left_relative_to_main_deck, 
                                                       max(main_deck_image.height, leader_base_image.height))
         leader_base_main_deck_image.paste(leader_base_image, 
                            (0, int(leader_base_main_deck_image.height / 2 - leader_base_image.height/ 2)), 
                            leader_base_image)
         leader_base_main_deck_image.paste(main_deck_image, 
-                           (leader_base_image.width + self.MAIN_DECK_COLUMN_SPACING, int(leader_base_main_deck_image.height / 2 - leader_base_main_deck_image.height/ 2)), 
+                           (leader_base_image.width + self._deck_list_image_generator_styles.leader_base_spacing_left_relative_to_main_deck, int(leader_base_main_deck_image.height / 2 - leader_base_main_deck_image.height/ 2)), 
                            main_deck_image)
         
-        if not self._is_sideboard_enabled:
+        if not self._deck_list_image_generator_styles.is_sideboard_enabled:
             sideboard_image = self._create_transparent_image(0, 0)
             sideboard_spacing = 0
         else:
             sideboard_image = self._create_sideboard_stacked_vertical_image(uniform_card_width, uniform_card_height)
-            sideboard_spacing = self.SIDEBOARD_LEFT_SPACING
+            sideboard_spacing = self._deck_list_image_generator_styles.sideboard_left_spacing_relative_to_main_deck
         result_image = self._create_transparent_image(leader_base_main_deck_image.width + sideboard_image.width + sideboard_spacing, 
                                                       max(leader_base_main_deck_image.height, sideboard_image.height))
         
@@ -251,12 +231,12 @@ class DraftListImageGenerator:
         uniform_card_width, uniform_card_height = self._uniform_card_dimensions()
         main_deck_image = self._create_main_deck_cost_curve_image(uniform_card_width, uniform_card_height)
         
-        if not self._is_sideboard_enabled:
+        if not self._deck_list_image_generator_styles.is_sideboard_enabled:
             sideboard_image = self._create_transparent_image(0, 0)
             sideboard_spacing = 0
         else:
             sideboard_image = self._create_sideboard_stacked_vertical_image(uniform_card_width, uniform_card_height)
-            sideboard_spacing = self.SIDEBOARD_LEFT_SPACING
+            sideboard_spacing = self._deck_list_image_generator_styles.sideboard_left_spacing_relative_to_main_deck
 
         main_deck_sideboard_image = self._create_transparent_image(main_deck_image.width + sideboard_image.width + sideboard_spacing, 
                                                       max(main_deck_image.height, sideboard_image.height))
@@ -275,12 +255,12 @@ class DraftListImageGenerator:
                                                                               uniform_card_width)
 
         result_image = self._create_transparent_image(max(main_deck_sideboard_image.width, leader_base_image.width), 
-                                                      main_deck_sideboard_image.height + leader_base_image.height + self.MAIN_DECK_ROWS_SPACING)
+                                                      main_deck_sideboard_image.height + leader_base_image.height + self._deck_list_image_generator_styles.main_deck_row_spacing)
         result_image.paste(leader_base_image, 
                            (int(result_image.width / 2 - leader_base_image.width/ 2), 0), 
                            leader_base_image)
         result_image.paste(main_deck_sideboard_image, 
-                           (int(result_image.width / 2 - main_deck_sideboard_image.width/ 2), leader_base_image.height + self.MAIN_DECK_ROWS_SPACING), 
+                           (int(result_image.width / 2 - main_deck_sideboard_image.width/ 2), leader_base_image.height + self._deck_list_image_generator_styles.main_deck_row_spacing), 
                            main_deck_sideboard_image)
 
         return result_image